Corrected element.szTitle, strcpy, commented out QUOTE and DELETE
The main change was removing a memset that erased element.szTitle, switching strcpy to be (element.szTitle, pszInput) vs the other way around, and commenting out all of QUOTE and DELETE for now because 5.2 doesn't need them. The only thing I notice is the printing format is a little off and the costs might not be all there, but I'll make sure my driver is passing all the data correctly for that at least.