strtok()...
Updated Option and Value statements to copy the szTitle from pszInput, but only after pszInput was edited by strtok() to remove the trailing '\n' char at the end of the remaining buffer string. The reason being is that the trailing '\n' may cause unwanted formatting in printing the price menu, so instead it is better in my opinion for the programmer of printPriceMenu and printOne to place a '\n' in those functions when they actually want one.